+++ changelog.md
@@ -5,6 +5,7 @@
 * Output window now clears on each Run, Compile or Build
 * Updated UDF file location for functions moved in latest AutoIt update (v3.3.14.3) 
 * Added configuration option to hide variables when using Ctrl+Shift+O to view symbols
+* Implemented Workspace/Folder-wide Symbol Search (`Ctrl+T`)
 
 ## 0.1.8
 * Improved function signature helpers